feat: implement APA citation style
- Create ApaCitationTable implementation of GenericCitationTable
- Add formatting logic for single, two, and multiple authors
- Define APA-specific citation separators
- Support APA standard parenthetical citation format
